clippy

+++ libafl_qemu/src/modules/drcov.rs
@@ -168,7 +168,7 @@ impl<F> DrCovModule<F> {
                             continue 'pcs_full;
                         }
                         if *idm == *id {
-                            #[allow(clippy::unnecessary_cast)]
+                            #[allow(clippy::unnecessary_cast)] // for GuestAddr -> u64
                             match lengths.get(pc) {
                                 Some(block_length) => {
                                     drcov_vec.push(DrCovBasicBlock::new(
@@ -216,7 +216,8 @@ impl<F> DrCovModule<F> {
                     if !module_found {
                         continue 'pcs;
                     }
-                    #[allow(clippy::unnecessary_cast)]
+
+                    #[allow(clippy::unnecessary_cast)] // for GuestAddr -> u64
                     match lengths.get(pc) {
                         Some(block_length) => {
                             drcov_vec.push(DrCovBasicBlock::new(
@@ -286,11 +287,12 @@ where
 
             let mut module_mapping: RangeMap<u64, (u16, String)> = RangeMap::new();
 
+            #[allow(clippy::unnecessary_cast)] // for GuestAddr -> u64
             for (i, (r, p)) in qemu
                 .mappings()
                 .filter_map(|m| {
                     m.path()
-                        .map(|p| ((m.start())..(m.end()), p.to_string()))
+                        .map(|p| ((m.start() as u64)..(m.end() as u64), p.to_string()))
                         .filter(|(_, p)| !p.is_empty())
                 })
                 .enumerate()
